+++ Infrastructure/WarehouseRepository.cs
@@ -1,4 +1,4 @@
-﻿using Application.DTOs;
+﻿using Domain.DTOs;
 using Application.Interfaces;
 using Domain;
 
@@ -14,8 +14,13 @@ namespace Infrastructure
             _context = context ?? throw new ArgumentNullException(nameof(context));
 
         }
-        public List<Warehouse> GetAll()
+        public List<Warehouse> GetAll(int userId)
         {
+            UserInWarehouse uiw = _context.UserWarehouseTable.Where(u => u.UserId == userId).First() ?? throw new KeyNotFoundException();
+            if (uiw.AccessLevel != 4)
+            {
+                throw new Exception("Access denied");
+            }
             return _context.WarehouseTable.ToList();
         }
 
@@ -23,22 +28,37 @@ namespace Infrastructure
         {
             return _context.WarehouseTable.Find(id) ?? throw new KeyNotFoundException();
         }
-        public Warehouse Create(Warehouse warehouse)
+        public Warehouse Create(Warehouse warehouse, int userId)
         {
+            UserInWarehouse uiw = _context.UserWarehouseTable.Where(u => u.UserId == userId).First() ?? throw new KeyNotFoundException();
+            if (uiw.AccessLevel != 4)
+            {
+                throw new Exception("Access denied");
+            }
             _context.WarehouseTable.Add(warehouse);
             _context.SaveChanges();
             return warehouse;
         }
 
-        public Warehouse Update(Warehouse warehouse)
+        public Warehouse Update(Warehouse warehouse, int userId)
         {
+            UserInWarehouse uiw = _context.UserWarehouseTable.Where(u => u.UserId == userId).First() ?? throw new KeyNotFoundException();
+            if (uiw.AccessLevel != 4)
+            {
+                throw new Exception("Access denied");
+            }
             _context.WarehouseTable.Update(warehouse);
             _context.SaveChanges();
             return warehouse;
         }
 
-        public Warehouse Delete(int id)
+        public Warehouse Delete(int id, int userId)
         {
+            UserInWarehouse uiw = _context.UserWarehouseTable.Where(u => u.UserId == userId).First() ?? throw new KeyNotFoundException();
+            if (uiw.AccessLevel != 4)
+            {
+                throw new Exception("Access denied");
+            }
             var warehouseToDelete = _context.WarehouseTable.Find(id) ?? throw new KeyNotFoundException();
             _context.WarehouseTable.Remove(warehouseToDelete);
             _context.SaveChanges();
@@ -58,9 +78,14 @@ namespace Infrastructure
             return products;
 
         }
-        public Product CreateProduct(int warehouseId,Product product)
+        public Product CreateProduct(int warehouseId,Product product, int userId)
         {
-            
+            UserInWarehouse uiw = _context.UserWarehouseTable.Where(u => u.UserId == userId).First() ?? throw new KeyNotFoundException();
+            if (uiw.AccessLevel != 4)
+            {
+                throw new Exception("Access denied");
+            }
+
             _context.ProductTable.Add(product);
             _context.SaveChanges();
             int productId  = _context.ProductTable.Where(p => p.Name == product.Name && p.CompanyId == product.CompanyId).Select(p => p.Id).FirstOrDefault();
@@ -69,16 +94,26 @@ namespace Infrastructure
             return _context.ProductTable.Find(productId) ?? throw new KeyNotFoundException();
         }
 
-        public Product UpdateProduct(int warehouseId, Product product)
+        public Product UpdateProduct(int warehouseId, Product product, int userId)
         {
-            
+            UserInWarehouse uiw = _context.UserWarehouseTable.Where(u => u.UserId == userId).First() ?? throw new KeyNotFoundException();
+            if (uiw.AccessLevel != 4)
+            {
+                throw new Exception("Access denied");
+            }
+
             _context.ProductWarehouseTable.Update(new ProductInWarehouse { ProductId = product.Id, WarehouseId = warehouseId, Quantity = product.Quantity });
             _context.SaveChanges();
             return _context.ProductTable.Find(product.Id) ?? throw new KeyNotFoundException();
         }
 
-        public Product DeleteProduct(int warehouseId, int productId, bool deleteFromProductTable)
+        public Product DeleteProduct(int warehouseId, int productId, bool deleteFromProductTable, int userId)
         {
+            UserInWarehouse uiw = _context.UserWarehouseTable.Where(u => u.UserId == userId).First() ?? throw new KeyNotFoundException();
+            if (uiw.AccessLevel != 4)
+            {
+                throw new Exception("Access denied");
+            }
             ProductInWarehouse pin = _context.ProductWarehouseTable.Where(p => p.ProductId == productId && p.WarehouseId == warehouseId).FirstOrDefault() ?? throw new KeyNotFoundException();
             _context.ProductWarehouseTable.Remove(pin);
             Product productToDelete = _context.ProductTable.Find(productId) ?? throw new KeyNotFoundException();
@@ -90,8 +125,13 @@ namespace Infrastructure
             return new Product() { Name = productToDelete.Name, Unit = productToDelete.Unit, PricePerUnit = productToDelete.PricePerUnit, CompanyId = productToDelete.CompanyId, Quantity = pin.Quantity };
         }
 
-        public Product AddProduct(ProductInWarehouse pin)
+        public Product AddProduct(ProductInWarehouse pin, int userId)
         {
+            UserInWarehouse uiw = _context.UserWarehouseTable.Where(u => u.UserId == userId).First() ?? throw new KeyNotFoundException();
+            if (uiw.AccessLevel != 4)
+            {
+                throw new Exception("Access denied");
+            }
             _context.ProductWarehouseTable.Add(pin);
             _context.SaveChanges();
             Product product = _context.ProductTable.Find(pin.ProductId) ?? throw new KeyNotFoundException();
@@ -110,8 +150,13 @@ namespace Infrastructure
             return users;
         }
 
-        public User AddUser(UserInWarehouse uiw)
+        public User AddUser(UserInWarehouse uiw, int userId)
         {
+            UserInWarehouse requestUiw = _context.UserWarehouseTable.Where(u => u.UserId == userId).First() ?? throw new KeyNotFoundException();
+            if (uiw.AccessLevel != 4)
+            {
+                throw new Exception("Access denied");
+            }
             _context.UserWarehouseTable.Add(uiw);
             _context.SaveChanges();
             return _context.UserTable.Find(uiw.UserId) ?? throw new KeyNotFoundException();
@@ -127,8 +172,13 @@ namespace Infrastructure
             return user;
         }
 
-        public User UpdateUserAccesLevel(UserInWarehouse uiw)
+        public User UpdateUserAccesLevel(UserInWarehouse uiw, int userId)
         {
+            UserInWarehouse requestUiw = _context.UserWarehouseTable.Where(u => u.UserId == userId).First() ?? throw new KeyNotFoundException();
+            if (requestUiw.AccessLevel != 4)
+            {
+                throw new Exception("Access denied");
+            }
             _context.UserWarehouseTable.Update(uiw);
             _context.SaveChanges();
             return _context.UserTable.Find(uiw.UserId) ?? throw new KeyNotFoundException();
